fix bgm property int overflow issue

+++ Sm5sh.GUI/ViewModels/ReactiveObjects/BgmEntryEditableViewModel.cs
@@ -61,13 +61,13 @@ namespace Sm5sh.GUI.ViewModels
         {
             public BgmEntryEditableViewModel Parent { get; }
             public string NameId { get { return Parent.ToneId; } }
-            public ulong LoopStartMs { get; set; }
-            public ulong LoopStartSample { get; set; }
-            public ulong LoopEndMs { get; set; }
-            public ulong LoopEndSample { get; set; }
-            public ulong TotalTimeMs { get; set; }
-            public ulong TotalSamples { get; set; }
-            public ulong Frequency { get { return TotalTimeMs / 1000 * TotalSamples; } }
+            public uint LoopStartMs { get; set; }
+            public uint LoopStartSample { get; set; }
+            public uint LoopEndMs { get; set; }
+            public uint LoopEndSample { get; set; }
+            public uint TotalTimeMs { get; set; }
+            public uint TotalSamples { get; set; }
+            public uint Frequency { get { return TotalTimeMs / 1000 * TotalSamples; } }
 
             public BgmPropertyEntryEditableViewModel(BgmEntryEditableViewModel parent)
             {
