+++ Sm5sh/Mods/Sm5sh.Mods.Music/Models/AudioCuePoints.cs
@@ -2,12 +2,12 @@
 {
     public class AudioCuePoints
     {
-        public ulong LoopStartMs { get; set; }
-        public ulong LoopStartSample { get; set; }
-        public ulong LoopEndMs { get; set; }
-        public ulong LoopEndSample { get; set; }
-        public ulong TotalTimeMs { get; set; }
-        public ulong TotalSamples { get; set; }
-        public ulong Frequency { get { return TotalTimeMs / 1000 * TotalSamples; } }
+        public uint LoopStartMs { get; set; }
+        public uint LoopStartSample { get; set; }
+        public uint LoopEndMs { get; set; }
+        public uint LoopEndSample { get; set; }
+        public uint TotalTimeMs { get; set; }
+        public uint TotalSamples { get; set; }
+        public uint Frequency { get { return TotalTimeMs / 1000 * TotalSamples; } }
     }
 }
